Rewrite and fix old STLC implementation
This is a refactored and improved version (it fixes all the bugs and is more featureful) of an old STLC I wrote before I started my Pie implementation. I decided to rewrite it and publish it because it's a good minimum viable product of a statically typed pure functional language.